+++ src/components/FlexibleForm/ControlDecorators/partials/index.js
@@ -1,2 +1,1 @@
-export * from './TextDate';
 export * from './OptionsList';