+++ src/routes/ViewAllLogs/ViewAllLogs.js
@@ -39,6 +39,7 @@ const {
 
 const columnMapping = {
   fileName: <FormattedMessage id="ui-data-import.fileName" />,
+  status: <FormattedMessage id="ui-data-import.status" />,
   hrId: <FormattedMessage id="ui-data-import.jobExecutionHrId" />,
   jobProfileName: <FormattedMessage id="ui-data-import.jobProfileName" />,
   totalRecords: <FormattedMessage id="ui-data-import.records" />,
@@ -48,11 +49,12 @@ const columnMapping = {
 
 const visibleColumns = [
   'fileName',
-  'hrId',
-  'jobProfileName',
+  'status',
   'totalRecords',
+  'jobProfileName',
   'completedDate',
   'runBy',
+  'hrId',
 ];
 
 const INITIAL_RESULT_COUNT = 25;
@@ -98,6 +100,7 @@ class ViewAllLogs extends Component {
             logsSearchTemplate,
             {
               fileName: 'fileName',
+              status: 'status',
               hrId: 'hrId/number',
               jobProfileName: 'jobProfileInfo.name',
               totalRecords: 'progress.total/number',
@@ -187,6 +190,7 @@ class ViewAllLogs extends Component {
           {record.fileName}
         </Button>
       ),
+      status: listTemplate({ intl }).status,
       runBy: listTemplate({ intl }).runBy,
       completedDate: listTemplate({ intl }).completedDate,
       jobProfileName: listTemplate({ intl }).jobProfileName,