+++ src/settings/MappingProfiles/ViewMappingProfile.js
@@ -54,6 +54,8 @@ import {
   getEntity,
   getEntityTags,
   MARC_TYPES,
+  FIELD_MAPPINGS_FOR_MARC,
+  FIELD_MAPPINGS_FOR_MARC_OPTIONS,
 } from '../../utils';
 
 import sharedCss from '../../shared.css';
@@ -205,6 +207,7 @@ export class ViewMappingProfile extends Component {
       incomingRecordType,
       existingRecordType,
       mappingDetails,
+      mappingDetails: { marcMappingOption },
     } = mappingProfile;
 
     const associations = [
@@ -214,11 +217,28 @@ export class ViewMappingProfile extends Component {
 
     const isMARCRecord = existingRecordType === MARC_TYPES.MARC_BIBLIOGRAPHIC;
 
+    const marcMappingOptionLabel = FIELD_MAPPINGS_FOR_MARC_OPTIONS.find(option => option.value === marcMappingOption)?.label;
+
+    const renderMARCTableView = () => {
+      const defaultFieldMappingForMARCColumns = ['action', 'field', 'indicator1', 'indicator2',
+        'subfield', 'subaction', 'data', 'position'];
+      const updatesFieldMappingForMARCColumns = ['field', 'indicator1', 'indicator2', 'subfield'];
+
+      const fieldMappingForMARCColumns = { [FIELD_MAPPINGS_FOR_MARC.UPDATES]: updatesFieldMappingForMARCColumns };
+
+      return (
+        <MARCTableView
+          columns={fieldMappingForMARCColumns[marcMappingOption] || defaultFieldMappingForMARCColumns}
+          fields={mappingDetails?.marcMappingDetails}
+        />
+      );
+    };
+
     const renderDetails = {
       INSTANCE: <MappingInstanceDetails mappingDetails={mappingDetails?.mappingFields} />,
       HOLDINGS: <MappingHoldingsDetails mappingDetails={mappingDetails?.mappingFields} />,
       ITEM: <MappingItemDetails mappingDetails={mappingDetails?.mappingFields} />,
-      MARC_BIBLIOGRAPHIC: <MARCTableView fields={mappingDetails?.marcMappingDetails} />,
+      MARC_BIBLIOGRAPHIC: renderMARCTableView(),
     };
 
     return (
@@ -262,6 +282,13 @@ export class ViewMappingProfile extends Component {
                 <FormattedMessage id={FOLIO_RECORD_TYPES[existingRecordType].captionId} />
               </div>
             </KeyValue>
+            {isMARCRecord && (
+              <KeyValue label={<FormattedMessage id="ui-data-import.fieldMappingsForMarc" />}>
+                <div data-test-field-mapping-for-marc-field>
+                  <FormattedMessage id={marcMappingOptionLabel} />
+                </div>
+              </KeyValue>
+            )}
             <KeyValue label={<FormattedMessage id="ui-data-import.description" />}>
               <div data-test-description>{mappingProfile.description || <NoValue />}</div>
             </KeyValue>
@@ -288,6 +315,7 @@ export class ViewMappingProfile extends Component {
                     <MappedHeader
                       mappedLabelId="ui-data-import.settings.profiles.select.mappingProfiles"
                       mappableLabelId={MAPPING_DETAILS_HEADLINE[existingRecordType]?.labelId}
+                      mappingTypeLabelId={marcMappingOptionLabel}
                       headlineProps={{ margin: 'small' }}
                     />
                   </Col>