+++ src/utils/index.js
@@ -33,3 +33,4 @@ export * from './useForceUpdate';
 export * from './getDecoratorValue';
 export * from './updateValueWithTemplate';
 export * from './fillEmptyFieldsWithValue';
+export * from './getTrimmedValue';